ValorLiquido e validações

+++ CalculoCDB.API/Controllers/CalculoCdbController.cs
@@ -13,22 +13,22 @@ namespace CalculoCDB.API.Controllers
     {
         [System.Web.Mvc.HttpGet]
         [System.Web.Http.Cors.EnableCors(origins: "*", headers: "*", methods: "*")]
-        public async Task<ActionResult<IEnumerable<Calculo>>> Index(decimal? valor_inicial, int? num_meses)
+        public async Task<ActionResult<IEnumerable<Calculo>>> Index(double? valor_inicial, int? num_meses)
         {
 
             return CalculoCDB(valor_inicial, num_meses).ToList();
 
         }
 
-        public static List<Calculo> CalculoCDB(decimal? valor_inicial, int? num_meses)
+        public static List<Calculo> CalculoCDB(double? valor_inicial, int? num_meses)
         {
             List<Calculo> mesesCalculoCDB = new List<Calculo>();
 
             if (valor_inicial.HasValue)
             {
-                decimal valor_final = (decimal)valor_inicial;
-                decimal cdi = ValorCDI();
-                decimal tb = ValorTB();
+                double valor_final = (double)valor_inicial;
+                double cdi = ValorCDI();
+                double tb = ValorTB();
 
                 Calculo mesinicio = new Calculo();
                 mesinicio.Mes = 0;
@@ -48,15 +48,15 @@ namespace CalculoCDB.API.Controllers
             return mesesCalculoCDB;
         }
 
-        private static decimal ValorCDI()
+        private static double ValorCDI()
         {
-            decimal cdi = 0.9M / 100;
+            double cdi = 0.9 / 100;
             return cdi;
         }
 
-        private static decimal ValorTB()
+        private static double ValorTB()
         {
-            decimal tb = 108 / 100;
+            double tb = 108 / 100;
             return tb;
         }
     }
